A new approach to fitting: build up peaks rather than fit all
It seems that if we try incrementally adding peaks alternating with fitting
we can get better results.  The quality of the result is rather dependent
on the quality of each initial peak, which remains a bit of a challenge.

If this approach works going forward there is a lot of code we can now
trim.